Move file hashes to contracts
