+++ source/Calamari.Contracts/ArgoCD/TrackedSourceDetail.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Octopus.Calamari.Contracts.ArgoCD;
+
+public record TrackedSourceDetail(
+    string? CommitSha,
+    DateTimeOffset? CommitTimestamp,
+    int SourceIndex,
+    List<FileHash> ReplacedFiles,
+    List<FileJsonPatch> PatchedFiles);
