+++ source/Calamari/ArgoCD/Conventions/UpdateImageTag/FileUpdateResult.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Octopus.Calamari.Contracts.ArgoCD;
 
 namespace Calamari.ArgoCD.Conventions.UpdateImageTag;
