+++ source/Calamari/ArgoCD/ProcessApplicationResult.cs
@@ -3,20 +3,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Calamari.ArgoCD.Models;
+using Octopus.Calamari.Contracts.ArgoCD;
 
 namespace Calamari.ArgoCD
 {
-    public record FileHash(string FilePath, string Hash);
-
-    public record FileJsonPatch(string FilePath, string JsonPatch);
-
-    public record TrackedSourceDetail(
-        string? CommitSha,
-        DateTimeOffset? CommitTimestamp,
-        int SourceIndex,
-        List<FileHash> ReplacedFiles,
-        List<FileJsonPatch> PatchedFiles);
-
     public class ProcessApplicationResult(
         string gatewayId,
         ApplicationName applicationName,
